refactor: update imports to use namespaces for built-in AI functionality
