+++ src/app/drawers/settings/AISettings.tsx
@@ -7,7 +7,10 @@ import ListItemText from "@mui/material/ListItemText";
 import Stack from "@mui/material/Stack";
 import TextField from "@mui/material/TextField";
 import Typography from "@mui/material/Typography";
-import { type BuiltInAIName, isSupported } from "@shared/ai/built-in-ai/spec";
+import {
+  type BuiltInAIName,
+  isSupported,
+} from "@shared/ai/built-in-ai/namespaces";
 import { useSharedContext } from "@shared/ai/sharedContext";
 
 const AI_FEATURES = [
