+++ src/app/drawers/settings/LanguageSettings.tsx
@@ -7,7 +7,7 @@ import MenuItem from "@mui/material/MenuItem";
 import Select from "@mui/material/Select";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
-import { isSupported } from "@shared/ai/built-in-ai/spec";
+import { isSupported } from "@shared/ai/built-in-ai/namespaces";
 import { useDownloadProgress } from "@shared/ai/downloadProgress";
 import { useLanguage } from "@shared/language/useLanguage";
