+++ src/shared/ai/built-in-ai/namespaces.ts
